grpc-api/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0